+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: jchene <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/08 14:26:28 by jchene            #+#    #+#             */
-/*   Updated: 2019/11/13 12:51:45 by jchene           ###   ########.fr       */
+/*   Updated: 2019/11/14 13:00:11 by jchene           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,12 +20,12 @@ size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize)
 	if (!(src))
 		return (0);
 	if (dstsize == 0)
-		return ((unsigned int)ft_strlen((char *)src));
+		return (ft_strlen((char *)src));
 	while (i < dstsize - 1 && src[i])
 	{
 		dst[i] = src[i];
 		i++;
 	}
 	dst[i] = '\0';
-	return ((unsigned int)ft_strlen((char *)src));
+	return (ft_strlen((char *)src));
 }
